requirements.txt
streamlit
python-dotenv
aiohttp
openai
azure-search-documents==11.6.0b4